discord-bot/requirements.txt
discord.py>=2.4.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
